refactor(transpiler): code reorg
